# refactor(selector-generator): Polish the code