… Cargo.lock
… etc/selector-generator-data/selectors.json